zeroes8.c: the state deteriorates due to non-invertible mixing
The ZrHa construction uses non-invertible mixing, which is generally
considered a bad thing.  To see how bad it can get in practice, I
downscale the construction by a factor of 4 (so that it uses 8x8->16
multiplication).

The hashing process occurs in two phases: feed and mix.  After a portion
of input data is fed into the internal state (usually with XOR or ADD),
the state needs to be "mixed": we need to disrupt the state and amplify
the effect of each input bit, so that the next portion of input data
cannot cancel out with the previous portion.  Mixing is said to be
invertible if it can be undone.  For example, the following mixing
step, with the state variables x and y, is invertible:

    m = x * C;
    y += m;

It can be undone:

    m = x * C;
    y -= m;

But the following mixing step, resembling the ZrHa construction, cannot
be undone:

    m0 = x * C;
    m1 = y * C;
    y += m0;
    x += m1;

It "ties the knot", it's no longer possible to figure out what needs
to be subtracted from x and y.  To keep it invertible, we would need
to execute the two steps successively:

    m0 = x * C;
    y += m0;
    m1 = y * C;
    x += m1;

But this results in a less efficient sequence of instructions, because
multiplications m0 and m1 cannot execute in parallel.  Invertible mixing
can execute more efficiently only if we have at least 3 state variables.

The fact that the mixing step cannot be undone actually means that the
mapping from the original state to the mixed state is not 1-to-1 (not a
permutation).  All inputs are possible, but some outputs are not
possible, so a few inputs must map to the same output.  Roughly speaking,
it loses data.  This is why non-invertible mixing is usually frowned
upon as contributing to collisions.

However, it may or may not be all that bad in practice.  The best case
is, after the state is mixed and cannot assume certain values, so what,
after feeding the next portion of data we may reasonably claim that the
state can assume any possible value again, more or less uniformly.  The
worst case though is that the state can deteriorate progressively,
assuming fewer and fewer possible values after each mixing steps.
This happens when you feed in zeroes.  Eventually the state "collapses":
one of the multipliers becomes zero, and the state alternates between
two (or sometimes four) possible values, as in this example:

    3500 2d00 bd00 bb00
    002d 0035 00bb 00bd
    3500 2d00 bd00 bb00
    002d 0035 00bb 00bd

But if we look at it from another angle, non-invertible mixing has a
plus side.  For one, you cannot recover the seed (other than by brute
force), and so you shouldn't be able to generate collisions too easily.
Bob Jenkins worried about the mixing functions that can "spontaneously
unmix".  Well, "tying the knot" will never unmix.  So it all depends
on how fast the state can deteriorate in practice.

The key realization I had a few days ago is that the rate of
deterioration actually does not depend (exclusively) on how wide
multiplication is.  Rather it depends on how big the whole state is.
In other words, if we update SIMD lanes/registers independently, we
expose ourselves to much higher deterioration than necessary.  If
instead we intermix the state among the two SIMD registers (which comes
for free in practice), the number of zero-updates before the state
collapses improves from 2^k to at least 2^(2k).

$ gcc -O2 zeroes8.c && ./a.out A |head -1024 |cut -f2 |stats.pl
q25=587
q50=1482
q75=3117
$ gcc -O2 zeroes8.c && ./a.out B |head -256 |cut -f2 |stats.pl
q25=20790430
q50=52553590
q75=98318214
$ gcc -O2 zeroes8.c && ./a.out C |head -256 |cut -f2 |stats.pl
q25=61713577
q50=181047618
q75=400268192