zeroes16.c: upscale and evaluate with 16x16->32 multiplication
With 16x16->32 multiplication, the original construction (updateA)
collapses after about 2^26 updates.

$ gcc -g -O2 -Wall zeroes16.c && ./a.out |head -256 |cut -f2 |stats.pl
q25=17629387
q50=55764622
q75=118887039

There is no point in trying to run simulations with the improved
constructions (updateB and updateC), because they must take at least
2^52, and our resource limited to 2^32 per try(), more or less.

Here is a rough estimate of how many zero-updates it takes for the
upscaled construction to collapse.

            updateA   updateB     udpateC
  8x8->16   2^11      2^25        2^27
16x16->32   2^26      2^53        2^56
32x32->64   2^57      2^115       2^119