collisions.c: replaced qsort.h with radix sort
Radix sort runs almost 2x faster (1.96x on Haswell)!  According to perf,
the share of hsort() in the running time goes down from 50%+ to 35%, so
this is a substantial improvement.  The only downside is that it can't
sort inplace.  In practice, this means that we have to allocate about
320M extra memory, while the string slab takes about 2G.  The algorithm
is cache-friendly and has good locality, so the extra memory requirement
does not translate into systematic delays due to cache misses.

Clever loop unrolling (the "n & 1" and "i += 2" pieces) helps only
a little, by 1-2%.  I also tried to reduce the number of radix loops
from 8 to 6, with 11-bit digits instead of 8-bit digits.  It ran slower.