hash8.h: initial revision
So this is where I left off in January.  The construction in hash8.h
is scaled down twice, from 32x32->64 down to 16x16->32 down to 8x8->16
multiplication (while uint16_t[2] mimics an xmm register).  With 8x8->16
multiplication, the probability of a multiplication "not going well at
all" (the worst case being multiplication by zero) is roughly 1/256.

Now that multiplication malfunctions often enough, we can pay proper
attention to the rest of our construction.  We try to minimize the
number of collisions by tweaking the construction and running empirical
tests.  For example, is it better to shuffle the state before the second
feeding of the data, or after?  We can just try a few options and see
which works best.  (It turns out that one xmm register should be
shuffled before, and the other after the second feeding of the data.)

Finally, what is the best way to feed and to combine?  We can simply
enumerate all feed/combine ops and try all the various combinations.
Something like this:

for F0 in Xor Add Sub; do
for F1 in Xor Add Sub; do
for F2 in Xor Add Sub; do
for F3 in Xor Add Sub; do
for F4 in Xor Add Sub; do
for F5 in Xor Add Sub; do
set -- -DF0=$F0 -DF1=$F1 -DF2=$F2 -DF3=$F3 -DF4=$F4 -DF5=$F5
...

I then do some layman's statistics and sort by the average number of
collisions.  So here are some of the best and some of the worst
combinations:

-DF0=Add -DF1=Xor -DF2=Add -DF3=Sub -DF4=Add -DF5=Sub avg=1.2 dev=2.09 q25=0 q50=0 q75=2.25
-DF0=Add -DF1=Xor -DF2=Add -DF3=Add -DF4=Add -DF5=Sub avg=1.5 dev=2.12 q25=0 q50=0 q75=3
-DF0=Add -DF1=Xor -DF2=Add -DF3=Sub -DF4=Sub -DF5=Xor avg=1.9 dev=2.33 q25=0 q50=1.5 q75=3
-DF0=Add -DF1=Sub -DF2=Xor -DF3=Add -DF4=Sub -DF5=Add avg=2.4 dev=2.75 q25=0 q50=3 q75=3
-DF0=Add -DF1=Xor -DF2=Add -DF3=Sub -DF4=Add -DF5=Add avg=2.4 dev=2.75 q25=0 q50=1.5 q75=5.25
-DF0=Add -DF1=Xor -DF2=Add -DF3=Add -DF4=Add -DF5=Add avg=2.7 dev=4.11 q25=0 q50=0 q75=5.25
-DF0=Add -DF1=Xor -DF2=Add -DF3=Add -DF4=Sub -DF5=Sub avg=3.3 dev=3.59 q25=0 q50=3 q75=5.25
-DF0=Add -DF1=Sub -DF2=Add -DF3=Sub -DF4=Xor -DF5=Add avg=3.9 dev=6.33 q25=0 q50=3 q75=3
-DF0=Add -DF1=Add -DF2=Xor -DF3=Add -DF4=Xor -DF5=Add avg=4.2 dev=2.89 q25=3 q50=4.5 q75=6
-DF0=Add -DF1=Xor -DF2=Sub -DF3=Sub -DF4=Sub -DF5=Add avg=4.8 dev=9.07 q25=0 q50=3 q75=3
...
-DF0=Add -DF1=Add -DF2=Xor -DF3=Xor -DF4=Xor -DF5=Xor avg=370.9 dev=447.76 q25=166.75 q50=209 q75=323.25
-DF0=Add -DF1=Xor -DF2=Sub -DF3=Xor -DF4=Xor -DF5=Sub avg=377.7 dev=404.61 q25=155.75 q50=208.5 q75=411
-DF0=Add -DF1=Xor -DF2=Xor -DF3=Xor -DF4=Sub -DF5=Add avg=394.5 dev=418.60 q25=198 q50=249.5 q75=323.5
-DF0=Add -DF1=Add -DF2=Xor -DF3=Xor -DF4=Sub -DF5=Sub avg=412.8 dev=374.82 q25=207.75 q50=237 q75=508.25
-DF0=Add -DF1=Xor -DF2=Sub -DF3=Xor -DF4=Xor -DF5=Xor avg=444.6 dev= 88.97 q25=410.75 q50=452 q75=493.75
-DF0=Add -DF1=Sub -DF2=Xor -DF3=Xor -DF4=Xor -DF5=Xor avg=455.1 dev=476.16 q25=179.75 q50=239 q75=388.5
-DF0=Add -DF1=Xor -DF2=Add -DF3=Xor -DF4=Xor -DF5=Xor avg=465.0 dev=119.71 q25=376.5 q50=444 q75=560.75
-DF0=Add -DF1=Xor -DF2=Xor -DF3=Xor -DF4=Xor -DF5=Xor avg=514.1 dev=70.041 q25=472.25 q50=513.5 q75=551.75
-DF0=Add -DF1=Xor -DF2=Xor -DF3=Xor -DF4=Add -DF5=Xor avg=562.6 dev=180.27 q25=443.5 q50=529.5 q75=604.25
-DF0=Add -DF1=Xor -DF2=Xor -DF3=Xor -DF4=Sub -DF5=Xor avg=636.0 dev=375.48 q25=492 q50=518.5 q75=542.75

It seems that there should be at least one Xor and at most two Xors, but
3 or 4 Xors are much worse.  I now need to make some changes to hash8.h
(to make it more similar to ../pkgelfsym/symhash.h) and rerun the empirical
tests a bit more rigorously.